pyproject.toml
[project]
name = "norde-maison"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "django-colorfield (>=0.14.0,<0.15.0)",
    "django-image-uploader-widget (>=1.1.0,<2.0.0)",
    "django-admin-sortable2 (>=2.3.1,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "faker (>=40.4.0,<41.0.0)",
    "django-tinymce (>=5.0.0,<6.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
